Reorganize irreducibility and primitivity tests
