+++ src/galois/_polys/__init__.py
@@ -8,3 +8,6 @@ from ._primitive import *
 
 # pylint: disable=undefined-variable
 _poly.GCD = _functions.gcd
+
+Poly.is_irreducible = _irreducible.is_irreducible
+Poly.is_primitive = _primitive.is_primitive
