+++ src/galois/_polys/_poly.py
@@ -12,7 +12,6 @@ from typing_extensions import Literal, Self
 
 from .._domains import Array, _factory
 from .._helper import export, verify_isinstance, verify_issubclass
-from .._prime import factors
 from ..typing import ArrayLike, ElementLike, PolyLike
 from . import _binary, _dense, _sparse
 from ._conversions import (
@@ -1233,179 +1232,6 @@ class Poly:
             return p_prime.derivative(k)
 
         return p_prime
-
-    def is_irreducible(self) -> bool:
-        r"""
-        Determines whether the polynomial :math:`f(x)` over :math:`\mathrm{GF}(p^m)` is irreducible.
-
-        .. info::
-
-            This is a method, not a property, to indicate this test is computationally expensive.
-
-        Returns:
-            `True` if the polynomial is irreducible.
-
-        See Also:
-            irreducible_poly, irreducible_polys
-
-        Notes:
-            A polynomial :math:`f(x) \in \mathrm{GF}(p^m)[x]` is *reducible* over :math:`\mathrm{GF}(p^m)` if it can
-            be represented as :math:`f(x) = g(x) h(x)` for some :math:`g(x), h(x) \in \mathrm{GF}(p^m)[x]` of strictly
-            lower degree. If :math:`f(x)` is not reducible, it is said to be *irreducible*. Since Galois fields are not
-            algebraically closed, such irreducible polynomials exist.
-
-            This function implements Rabin's irreducibility test. It says a degree-:math:`m` polynomial :math:`f(x)`
-            over :math:`\mathrm{GF}(q)` for prime power :math:`q` is irreducible if and only if
-            :math:`f(x)\ |\ (x^{q^m} - x)` and :math:`\textrm{gcd}(f(x),\ x^{q^{m_i}} - x) = 1` for
-            :math:`1 \le i \le k`, where :math:`m_i = m/p_i` for the :math:`k` prime divisors :math:`p_i` of :math:`m`.
-
-        References:
-            - Rabin, M. Probabilistic algorithms in finite fields. SIAM Journal on Computing (1980), 273-280.
-              https://apps.dtic.mil/sti/pdfs/ADA078416.pdf
-            - Gao, S. and Panarino, D. Tests and constructions of irreducible polynomials over finite fields.
-              https://www.math.clemson.edu/~sgao/papers/GP97a.pdf
-            - Section 4.5.1 from https://cacr.uwaterloo.ca/hac/about/chap4.pdf
-            - https://en.wikipedia.org/wiki/Factorization_of_polynomials_over_finite_fields
-
-        Examples:
-            .. ipython:: python
-
-                # Conway polynomials are always irreducible (and primitive)
-                f = galois.conway_poly(2, 5); f
-
-                # f(x) has no roots in GF(2), a necessary but not sufficient condition of being irreducible
-                f.roots()
-
-                f.is_irreducible()
-
-            .. ipython:: python
-
-                g = galois.irreducible_poly(2**4, 2, method="random"); g
-                h = galois.irreducible_poly(2**4, 3, method="random"); h
-                f = g * h; f
-
-                f.is_irreducible()
-        """
-        # pylint: disable=too-many-return-statements
-        if self.degree == 0:
-            # Over fields, f(x) = 0 is the zero element of GF(p^m)[x] and f(x) = c are the units of GF(p^m)[x].
-            # Both the zero element and the units are not irreducible over the polynomial ring GF(p^m)[x].
-            return False
-
-        if self.degree == 1:
-            # f(x) = x + a (even a = 0) in any Galois field is irreducible
-            return True
-
-        if self.coeffs[-1] == 0:
-            # g(x) = x can be factored, therefore it is not irreducible
-            return False
-
-        if self.field.order == 2 and self.nonzero_coeffs.size % 2 == 0:
-            # Polynomials over GF(2) with degree at least 2 and an even number of terms satisfy f(1) = 0, hence
-            # g(x) = x + 1 can be factored. Section 4.5.2 from https://cacr.uwaterloo.ca/hac/about/chap4.pdf.
-            return False
-
-        field = self.field
-        q = field.order
-        m = self.degree
-        x = Poly.Identity(field)
-
-        primes, _ = factors(m)
-        h0 = Poly.Identity(field)
-        n0 = 0
-        for ni in sorted([m // pi for pi in primes]):
-            # The GCD of f(x) and (x^(q^(m/pi)) - x) must be 1 for f(x) to be irreducible, where pi are the
-            # prime factors of m.
-            hi = pow(h0, q ** (ni - n0), self)
-            g = GCD(self, hi - x)
-            if g != 1:
-                return False
-            h0, n0 = hi, ni
-
-        # f(x) must divide (x^(q^m) - x) to be irreducible
-        h = pow(h0, q ** (m - n0), self)
-        g = (h - x) % self
-        if g != 0:
-            return False
-
-        return True
-
-    def is_primitive(self) -> bool:
-        r"""
-        Determines whether the polynomial :math:`f(x)` over :math:`\mathrm{GF}(q)` is primitive.
-
-        .. info::
-
-            This is a method, not a property, to indicate this test is computationally expensive.
-
-        Returns:
-            `True` if the polynomial is primitive.
-
-        See Also:
-            primitive_poly, primitive_polys, conway_poly, matlab_primitive_poly
-
-        Notes:
-            A degree-:math:`m` polynomial :math:`f(x)` over :math:`\mathrm{GF}(q)` is *primitive* if it is
-            irreducible and :math:`f(x)\ |\ (x^k - 1)` for :math:`k = q^m - 1` and no :math:`k` less than
-            :math:`q^m - 1`.
-
-        References:
-            - Algorithm 4.77 from https://cacr.uwaterloo.ca/hac/about/chap4.pdf
-
-        Examples:
-            All Conway polynomials are primitive.
-
-            .. ipython:: python
-
-                f = galois.conway_poly(2, 8); f
-                f.is_primitive()
-
-                f = galois.conway_poly(3, 5); f
-                f.is_primitive()
-
-            The irreducible polynomial of :math:`\mathrm{GF}(2^8)` for AES is not primitive.
-
-            .. ipython:: python
-
-                f = galois.Poly.Degrees([8, 4, 3, 1, 0]); f
-                f.is_irreducible()
-                f.is_primitive()
-        """
-        if self.degree == 0:
-            # Over fields, f(x) = 0 is the zero element of GF(p^m)[x] and f(x) = c are the units of GF(p^m)[x].
-            # Both the zero element and the units are not irreducible over the polynomial ring GF(p^m)[x], and
-            # therefore cannot be primitive.
-            return False
-
-        if self.field.order == 2 and self.degree == 1:
-            # There is only one primitive polynomial in GF(2)
-            return self == Poly([1, 1])
-
-        if self.coeffs[-1] == 0:
-            # A primitive polynomial cannot have zero constant term
-            # TODO: Why isn't f(x) = x primitive? It's irreducible and passes the primitivity tests.
-            return False
-
-        if not self.is_irreducible():
-            # A polynomial must be irreducible to be primitive
-            return False
-
-        field = self.field
-        q = field.order
-        m = self.degree
-        one = Poly.One(field)
-
-        primes, _ = factors(q**m - 1)
-        x = Poly.Identity(field)
-        for ki in sorted([(q**m - 1) // pi for pi in primes]):
-            # f(x) must not divide (x^((q^m - 1)/pi) - 1) for f(x) to be primitive, where pi are the prime factors
-            # of q**m - 1.
-            h = pow(x, ki, self)
-            g = (h - one) % self
-            if g == 0:
-                return False
-
-        return True
 
     def is_square_free(self) -> bool:
         r"""
