+++ src/galois/_polys/_primitive.py
@@ -10,7 +10,8 @@ from typing_extensions import Literal
 from .._databases import ConwayPolyDatabase
 from .._domains import _factory
 from .._helper import export, verify_isinstance
-from .._prime import is_prime, is_prime_power
+from .._prime import factors, is_prime, is_prime_power
+from ._irreducible import is_irreducible
 from ._poly import (
     Poly,
     _deterministic_search,
@@ -19,6 +20,85 @@ from ._poly import (
     _random_search,
     _random_search_fixed_terms,
 )
+
+
+# NOTE: This is a method of the Poly class. It will be added to the Poly class in `polys/__init__.py`.
+def is_primitive(self) -> bool:
+    r"""
+    Determines whether the polynomial :math:`f(x)` over :math:`\mathrm{GF}(q)` is primitive.
+
+    .. info::
+
+        This is a method, not a property, to indicate this test is computationally expensive.
+
+    Returns:
+        `True` if the polynomial is primitive.
+
+    See Also:
+        primitive_poly, primitive_polys, conway_poly, matlab_primitive_poly
+
+    Notes:
+        A degree-:math:`m` polynomial :math:`f(x)` over :math:`\mathrm{GF}(q)` is *primitive* if it is
+        irreducible and :math:`f(x)\ |\ (x^k - 1)` for :math:`k = q^m - 1` and no :math:`k` less than
+        :math:`q^m - 1`.
+
+    References:
+        - Algorithm 4.77 from https://cacr.uwaterloo.ca/hac/about/chap4.pdf
+
+    Examples:
+        All Conway polynomials are primitive.
+
+        .. ipython:: python
+
+            f = galois.conway_poly(2, 8); f
+            f.is_primitive()
+
+            f = galois.conway_poly(3, 5); f
+            f.is_primitive()
+
+        The irreducible polynomial of :math:`\mathrm{GF}(2^8)` for AES is not primitive.
+
+        .. ipython:: python
+
+            f = galois.Poly.Degrees([8, 4, 3, 1, 0]); f
+            f.is_irreducible()
+            f.is_primitive()
+    """
+    if self.degree == 0:
+        # Over fields, f(x) = 0 is the zero element of GF(p^m)[x] and f(x) = c are the units of GF(p^m)[x].
+        # Both the zero element and the units are not irreducible over the polynomial ring GF(p^m)[x], and
+        # therefore cannot be primitive.
+        return False
+
+    if self.field.order == 2 and self.degree == 1:
+        # There is only one primitive polynomial in GF(2)
+        return self == Poly([1, 1])
+
+    if self.coeffs[-1] == 0:
+        # A primitive polynomial cannot have zero constant term
+        # TODO: Why isn't f(x) = x primitive? It's irreducible and passes the primitivity tests.
+        return False
+
+    if not is_irreducible(self):
+        # A polynomial must be irreducible to be primitive
+        return False
+
+    field = self.field
+    q = field.order
+    m = self.degree
+    one = Poly.One(field)
+
+    primes, _ = factors(q**m - 1)
+    x = Poly.Identity(field)
+    for ki in sorted([(q**m - 1) // pi for pi in primes]):
+        # f(x) must not divide (x^((q^m - 1)/pi) - 1) for f(x) to be primitive, where pi are the prime factors
+        # of q**m - 1.
+        h = pow(x, ki, self)
+        g = (h - one) % self
+        if g == 0:
+            return False
+
+    return True
 
 
 @export
